test(examples): make log regression Bun-compatible
